New policy: private things has no XML comments.
Too much redundancy is too much.

All the code is there. It says what it does.

If comes the day when others say this project is not properly
documented... I'll take it as meaning that they want tutorials and
examples. Not a bunch of comments that say the same thing that the code
says and that they won't know from where to start reading anyway.