Added LocalizedTexts
Arleen is now able to load localized texts for languages and regions by
the same mechanics that configurations is lodaded.

Added a localized version of the salute, this should work as example.

Localized text (in similar fashion to the configuration) can be an
embeded resource or an external file, the external file takes
preference.

No merge of resources is done.